Move `loader` under `tapioca::runtime`
