+++ lib/tapioca/generators/dsl.rb
@@ -53,7 +53,7 @@ module Tapioca
 
         super(default_command: default_command, file_writer: file_writer)
 
-        @loader = T.let(nil, T.nilable(Loader))
+        @loader = T.let(nil, T.nilable(Runtime::Loader))
       end
 
       sig { override.void }
@@ -326,9 +326,9 @@ module Tapioca
         end.sort
       end
 
-      sig { returns(Loader) }
+      sig { returns(Runtime::Loader) }
       def loader
-        @loader ||= Loader.new
+        @loader ||= Runtime::Loader.new
       end
 
       sig { params(class_name: String).returns(String) }
