+++ lib/tapioca/internal.rb
@@ -2,9 +2,9 @@
 # frozen_string_literal: true
 
 require "tapioca"
-require "tapioca/loader"
 require "tapioca/sorbet_ext/generic_name_patch"
 require "tapioca/sorbet_ext/fixed_hash_patch"
+require "tapioca/runtime/loader"
 require "tapioca/runtime/generic_type_registry"
 require "tapioca/helpers/cli_helper"
 require "tapioca/helpers/config_helper"
